fix(*) remove circular dependency on the Istio submodule
The Istio submodule imports Kuma config definitions, and Kuma calls
functions from the Istio submodule. Break the circular dependency by
moving the config definitions into the Istio submodule, even though
they more properly belong in Kuma itself.

This makes it possible to run `go mod tidy` in the Istio submodule.

This fixes #2087.
