+++ pkg/transparentproxy/istio/go.mod
@@ -3,9 +3,11 @@ module github.com/kumahq/kuma/pkg/transparentproxy/istio
 go 1.16
 
 require (
+	github.com/miekg/dns v1.0.14
 	github.com/pkg/errors v0.9.1
-	github.com/spf13/cobra v1.2.1 // indirect
+	github.com/spf13/cobra v1.2.1
 	github.com/spf13/viper v1.8.1
-	// fetched with go get istio.io/istio/tools/istio-iptables@1.7.8
-	istio.io/istio v0.0.0-20210223230603-30e54dcb8a1c
+	golang.org/x/sys v0.0.0-20210510120138-977fb7262007
+	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
+	istio.io/pkg v0.0.0-20201202160453-b7f8c8c88ca3
 )
